% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AKME-package.R
\docType{package}
\name{AKME-package}
\alias{AKME}
\alias{AKME-package}
\title{AKME: Adjusted Kaplan-Meier Estimator for Right-Censored, Site-Weighted Data}
\description{
Computes an adjusted (weighted) Kaplan-Meier estimator of the survival function for right-censored concentration data sampled unevenly across sites, following Xie and Liu (2005). Provides the weighted log-rank test of Singh et al. (2014) for comparing two groups via permutation, bootstrap, or simulation, along with quantile estimation, Greenwood standard errors, and ggplot2-based plotting helpers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cwnoack/weighted_kaplan_meier}
  \item Report bugs at \url{https://github.com/cwnoack/weighted_kaplan_meier/issues}
}

}
\author{
\strong{Maintainer}: Clint Noack \email{clint.noack@gmail.com}

}
\keyword{internal}
