% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_weights.R
\name{calc_weights}
\alias{calc_weights}
\title{Calculate site-specific observational weights as 1/n}
\usage{
calc_weights(site_vector)
}
\arguments{
\item{site_vector}{A vector of site identifiers.}
}
\value{
A tibble with columns \code{Site} and \code{weight}.
}
\description{
For pooled data from many sites (or any equivalent sampling group),
calculate the weight \code{1 / n_k} of an individual observation from site \code{k}.
}
\examples{
sites <- sample(letters[1:10], 100, replace = TRUE,
                prob = 1:10 / sum(1:10))
calc_weights(sites)
}
