% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/calc_weights.R
\name{calc_weights}
\alias{calc_weights}
\title{Calculate the site-specific observational weights as 1/n}
\usage{
calc_weights(site_vector)
}
\arguments{
\item{site_vector}{A vector of site IDs}
}
\description{
For pooled data from many sites, or other equivalent sampling groups,
calculate the weight of an individual observation from that site.
}
\examples{
sites <- sample(letters[1:10],100,replace = T, prob = 1:10/sum(1:10))
weights <- calc_weights(sites)
}

