% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/G_rho.R
\name{G_rho}
\alias{G_rho}
\title{Log-rank test statistic G_rho}
\usage{
G_rho(ref_data, comp_data, rho = 1)
}
\arguments{
\item{ref_data}{data.frame with three columns: (1) measured concentration, (2) flag for nondetects [BDL = 1], (3) unique site identifier.
comp_data data.frame with three columns: (1) measured concentration, (2) flag for nondetects [BDL = 1], (3) unique site identifier.
rho Positive, real number input. This is the exponential argument of the weighting function as in Singh et al. (2014)}
}
\value{
G real number value of test statistic
}
\description{
Calculates the test statistic G_rho for comparison of two data.frames: one with
the "reference" data and one with the "comparison" data. Additional input is
the value of rho to be used. A value of 1 returns a statistic weighted
towards the largest observations (concentrations) by multiplying by a function
of the pooled survival estimator, while a value of 0 returns an unweighted statistic.
}

