% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/G_rho_hist.R
\name{G_rho_hist}
\alias{G_rho_hist}
\title{Plotting function for log-rank test.}
\usage{
G_rho_hist(log_rank_output)
}
\arguments{
\item{log_rank_ouput}{output from `log_rank` function, a list}
}
\description{
Takes the output from the log-rank test (a list) and plots a histogram of
the simulated/permuted/bootstrapped G* values along with the test value
and associated P-value.
}

