% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/group_km.R
\name{group_km}
\alias{group_km}
\title{AKME for grouped data}
\usage{
group_km(grouped_data)
}
\arguments{
\item{grouped_data}{data.frame with four columns: (1) measured concentration,
(2) flag for nondetects [BDL = 1],
 (3) unique site identifier,
  (4) dataset identifier.}
}
\value{
observed A data.frame containing the observed concentrations, the
  weighted "at-risk" and "events" at each concentration, and the survival
  estimator along with some intermediate data.
}
\description{
Calculates the AKME for grouped data. Requires the same input as `Surv_weighted`
but with an additional column: "Dataset". Output will be same as `Surv_weighted` but for each group.
}

