% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/grp_quantiles.R
\name{grp_quantiles}
\alias{grp_quantiles}
\title{Weighted quantiles for grouped data}
\usage{
grp_quantiles(grouped_data, percentiles = c(0.05, 0.1, 0.25, 0.5, 0.75, 0.9,
  0.95), type = "interp", sig.fig = 3)
}
\arguments{
\item{grouped_data}{data.frame with four columns: (1) measured concentration,
(2) flag for nondetects [BDL = 1],
 (3) unique site identifier,
  (4) dataset identifier.}

\item{percentiles}{vector of desired percentiles}

\item{type}{method for calculation of quantiles, either:
(1) 'interp' for interpolation between nearest points,
(2) 'PiR' for nearest value that does not exceed the percentile, or
(3) 'nearest' for nearest value (that may exceed desired percentile)}

\item{sig.fig}{desired number of significant figures for estimate.}
}
\value{
quantile_tbl data.frame containing the desired percentiles and the
order statistic estimates of those percentiles from each group in `grouped_data`
}
\description{
Calculates the AKME for grouped data and returns a table of desired quantiles.
 Requires the same input as `Surv_quanitle` but with an additional column:
  "Dataset". Output will be same as `Surv_quantile` but for each group.
}

