% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/Surv_quantile.R
\name{Surv_quantile}
\alias{Surv_quantile}
\title{Weighted quantiles for censored data}
\usage{
Surv_quantile(censored_data, percentiles = c(0.05, 0.1, 0.25, 0.5, 0.75, 0.9,
  0.95), type = "interp", sig.fig = 3)
}
\arguments{
\item{censored_data}{data.frame with three columns: (1) measured concentration,
(2) flag for nondetects [BDL = 1], and
 (3) unique site identifier}

\item{percentiles}{vector of desired percentiles}

\item{type}{method for calculation of quantiles, either:
(1) 'interp' for interpolation between nearest points,
(2) 'PiR' for nearest value that does not exceed the percentile, or
(3) 'nearest' for nearest value (that may exceed desired percentile)}

\item{sig.fig}{desired number of significant figures for estimate.}
}
\value{
perc_df data.frame containing the desired percentiles and the
order statistic estimates of those percentiles from `censored_data`
}
\description{
Calculates the AKME for censored data and returns a table of desired quantiles.
 Output contains the desired percentiles and the corresponding order statistics.
 Code will throw warnings if one or more of the desired percentiles is less than
 the fraction of censored data and may produce NA values for order statistics at
 those percentiles depending on the `type` argument to the function
 (i.e. `nearest` will not return NA but may return repeated values).
}

