% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/Surv_wconf.R
\name{Surv_wconf}
\alias{Surv_wconf}
\title{Apply Greenwood's formula for S.E. estimation of survival function}
\usage{
Surv_wconf(wKM_object, alpha = 0.05)
}
\description{
Apply Greenwood's formula for S.E. estimation of survival function
}

